6/16/25 minor visual update to Color Titlescreen
Added the new titlescreen for the color variant, i might modify it to have a proper background, but its cool as is. Merging both the OG titlescreen and the steam release title screen. I've had this idea for like 2 years and it's nice to finally get around to doing it, Im going to have to edit it upward so it doesn't get covered up by the menu, but it works for now! Finally, the first real update of 2025. I used Prodigle's converter for this, it's not quite right here to be honest, for some reason the darkest color and second darkest are flipped in the palette system? But hey! it works.