6/18/25 minor update Niko visual bug stair fix
I looked into the stairs and fixed that, The issue there was a script of old code which was switching to a special stairfix version of the lightbulb sprite, this was before the PLATFORMER  system had stairs so i was stuck macgyvering a solution that. But now i don't need to. In the wise words of Todd Howard, "it just works" should finally fix the fireplace dialogue issue next (i forgor 4 years ago to actually add the fireplace_lit dialogue. So it uses the same dialogue text whether or not it's lit.)