6/19/25 minor update dialogue fix
Added in the fire dialogue. Maybe i should consider optimizing the code for interactions next? since it TANKS the framerate on the original Gameboy. Probably moving that to Init and manually handling all edge cases would work? Because the stack system might mess with things, i dunno. That's for tommorow me to figure out.